+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# v1.4.2
 - Fast and dirty mode, to find quickly any path
 - May check new obstacles while searching or not
 - Check the final position in XYO mode